# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Trader API
  description: |
    AI-powered stock recommendation service backend API
    
    This API provides comprehensive trading recommendation services, market data, 
    portfolio management, and user authentication features.
    
    ## Features
    - User authentication and profile management
    - Real-time stock market data integration via Finnhub API
    - AI-powered trading recommendations with multiple strategies
    - Portfolio tracking and performance analytics
    - Subscription-based access tiers
    - Advanced trading strategies (Momentum, Reversal, Swing Trading, etc.)
    
    ## Authentication
    This API uses JWT token-based authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Authentication endpoints: 5 requests per minute
    - General API endpoints: 100 requests per minute
    - Market data endpoints: 200 requests per minute
    
    ## Subscription Tiers
    - **Basic**: 3 recommendations per day, Jesse Livermore strategy only
    - **Premium**: 50 recommendations per day, access to multiple strategies
    - **Professional**: Unlimited recommendations, all strategies, advanced analytics
    
  version: 1.0.0
  contact:
    name: Trader API Team
    email: support@trader-api.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
                - investmentStyle
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  example: securePassword123
                name:
                  type: string
                  example: John Doe
                investmentStyle:
                  type: string
                  enum: [conservative, moderate, aggressive]
                  example: moderate
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Retrieve current user's profile information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    $ref: '#/components/schemas/UserProfile'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Authentication
      summary: Update user profile
      description: Update user profile information
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                investment_style:
                  type: string
                  enum: [conservative, moderate, aggressive]
                risk_tolerance:
                  type: number
                  minimum: 1
                  maximum: 10
                notification_preferences:
                  type: object
                  properties:
                    email_recommendations:
                      type: boolean
                    push_notifications:
                      type: boolean
                    sms_alerts:
                      type: boolean
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  profile:
                    $ref: '#/components/schemas/UserProfile'

  # Market Data Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Error code for programmatic handling
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: Registration successful
        user:
          $ref: '#/components/schemas/User'
        session:
          $ref: '#/components/schemas/Session'
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        investment_style:
          type: string
          enum: [conservative, moderate, aggressive]
    Session:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        token_type:
          type: string
          example: Bearer
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        investment_style:
          type: string
          enum: [conservative, moderate, aggressive]
        risk_tolerance:
          type: number
          minimum: 1
          maximum: 10
        subscription_tier:
          type: string
          enum: [basic, premium, professional]
        notification_preferences:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...